peerstore: move the protobuf to pstoreds
