+++ p2p/host/peerstore/pstoreds/addr_book_gc.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/libp2p/go-libp2p/core/peer"
-	pb "github.com/libp2p/go-libp2p/p2p/host/peerstore/pb"
+	pb "github.com/libp2p/go-libp2p/p2p/host/peerstore/pstoreds/pb"
 
 	ds "github.com/ipfs/go-datastore"
 	"github.com/ipfs/go-datastore/query"
